finish cleaning the after-login file
