+++ src/App.js
@@ -1,8 +1,8 @@
 import { Route, BrowserRouter as Router, Routes } from "react-router-dom"
 import About from "./components/About/about"
-import Dashboard from "./components/After-login/Dashboard/dashboard"
-import MyTasks from "./components/After-login/My-tasks/my-tasks"
-import Profile from "./components/After-login/Profile/profile"
+import Dashboard from "./components/After-login/Dashboard/Dashboard"
+import MyTasks from "./components/After-login/MyTasks/MyTasks"
+import Profile from "./components/After-login/Profile/Profile"
 import Assessment from "./components/Assessment/assessment"
 import Question1 from "./components/Assessment/question1"
 import Question2 from "./components/Assessment/question2"
