+++ src/components/After-login/Dashboard/FilterIcons.js
@@ -1,0 +1,18 @@
+import careIcon from "../../../assets/Care.png";
+import careBlueIcon from "../../../assets/Care_blue.png";
+import cleanIcon from "../../../assets/Clean.png";
+import cleanBlueIcon from "../../../assets/Clean_blue.png";
+import cookIcon from "../../../assets/Cook.png";
+import cookBlueIcon from "../../../assets/Cook_blue.png";
+import emotionIcon from "../../../assets/Emotion.png";
+import emotionBlueIcon from "../../../assets/Emotion_blue.png";
+import repairIcon from "../../../assets/Repair.png";
+import repairBlueIcon from "../../../assets/Repair_blue.png";
+
+export const filterIcons = [
+  { name: "Cook", icon: cookIcon, activeIcon: cookBlueIcon },
+  { name: "Clean", icon: cleanIcon, activeIcon: cleanBlueIcon },
+  { name: "Care", icon: careIcon, activeIcon: careBlueIcon },
+  { name: "Emotion", icon: emotionIcon, activeIcon: emotionBlueIcon },
+  { name: "Repair", icon: repairIcon, activeIcon: repairBlueIcon },
+];
