+++ src/components/After-login/Navigation/NavItem.jsx
@@ -1,0 +1,19 @@
+import React from "react";
+import { Link, useLocation } from "react-router-dom";
+
+export default function NavItem({ to, icon, label }) {
+  const location = useLocation();
+  const isActive = location.pathname === to;
+
+  return (
+    <Link
+      to={to}
+      className={`flex items-center space-x-3 px-4 py-3 rounded-xl ${
+        isActive ? "bg-[#FFF8E7]" : "text-gray-600 hover:bg-gray-50"
+      }`}
+    >
+      {icon}
+      <span>{label}</span>
+    </Link>
+  );
+}
