+++ src/components/After-login/Navigation/SideBar.jsx
@@ -1,0 +1,76 @@
+import React from "react";
+import NavItem from "./NavItem";
+import pic from "../../../assets/pic.jpeg";
+
+export default function Sidebar() {
+  // Define navigation items with their icons
+  const navItems = [
+    {
+      to: "/dashboard",
+      label: "Dashboard",
+      icon: (
+        <svg className="w-5 h-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+          <path
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            strokeWidth={2}
+            d="M3 12l2-2m0 0l7-7 7 7M5 10v10a1 1 0 001 1h3m10-11l2 2m-2-2v10a1 1 0 01-1 1h-3m-6 0a1 1 0 001-1v-4a1 1 0 011-1h2a1 1 0 011 1v4a1 1 0 001 1m-6 0h6"
+          />
+        </svg>
+      )
+    },
+    {
+      to: "/my-tasks",
+      label: "My Tasks",
+      icon: (
+        <svg className="w-5 h-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+          <path
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            strokeWidth={2}
+            d="M3 12l2-2m0 0l7-7 7 7M5 10v10a1 1 0 001 1h3m10-11l2 2m-2-2v10a1 1 0 01-1 1h-3m-6 0a1 1 0 001-1v-4a1 1 0 011-1h2a1 1 0 011 1v4a1 1 0 001 1m-6 0h6"
+          />
+        </svg>
+      )
+    },
+    {
+      to: "/profile",
+      label: "Profile",
+      icon: (
+        <svg className="w-5 h-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+          <path
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            strokeWidth={2}
+            d="M16 7a4 4 0 11-8 0 4 4 0 018 0zM12 14a7 7 0 00-7 7h14a7 7 0 00-7-7z"
+          />
+        </svg>
+      )
+    }
+  ];
+
+  return (
+    <aside className="w-72 min-h-[calc(100vh-64px)] border-r border-gray-200 p-6">
+      {/* User Profile */}
+      <div className="flex items-center space-x-3 p-4 rounded-2xl border border-gray-100 mb-8 cursor-pointer">
+        <img src={pic || "/placeholder.svg"} alt="Profile" className="w-10 h-10 rounded-full" />
+        <div>
+          <h3 className="font-medium">Kath</h3>
+          <p className="text-sm text-gray-500">xiaoqz6@uw.edu</p>
+        </div>
+        <button className="ml-auto">
+          <svg className="w-5 h-5 text-gray-400" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 5l7 7-7 7" />
+          </svg>
+        </button>
+      </div>
+
+      {/* Navigation Menu */}
+      <nav className="space-y-2">
+        {navItems.map((item) => (
+          <NavItem key={item.to} to={item.to} icon={item.icon} label={item.label} />
+        ))}
+      </nav>
+    </aside>
+  );
+}
